Move generic query and records test to the queries package
